convert rostest to unittest
and reduce code duplication
