--- pilz_teleoperation/test/unittest_ros_message_serializer.test
@@ -1,27 +1,0 @@
-<!--
-Copyright (c) 2019 Pilz GmbH & Co. KG
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
-<launch>
-  <arg name="coverage" default="false"/>
-  <arg name="pythontest_launch_prefix" value="$(eval 'python-coverage run -p' if arg('coverage') else '')"/>
-
-  <!-- test node -->
-  <test test-name="test_message_serialization" pkg="pilz_teleoperation"
-    type="unittest_ros_message_serializer.py" time-limit="200"
-    launch-prefix="$(arg pythontest_launch_prefix)"/>
-
-</launch>
